pyproject.toml
[project]
name = "sqlpilot"
version = "0.1.0"
description = "Chatbot oriented to Text2Sql using llama-index and Snowflake with Streamlit"
authors = [
    {name = "Erick Calderin"}
]

readme = "README.md"
requires-python = "==3.11.0"
dependencies = [
    "llama-index-core (>=0.12.15,<0.13.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "snowflake-sqlalchemy (>=1.7.3,<2.0.0)",
    "streamlit (>=1.41.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "snowflake-connector-python[pandas] (>=3.13.2,<4.0.0)",
    "snowflake-snowpark-python (>=1.27.0,<2.0.0)",
    "llama-index-llms-openai (>=0.3.19,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.test.dependencies]
ruff = "^0.9.4"

[tool.poetry]
package-mode = false
